refactor: allow injecting registry into getters
