+++ src/services/onMessage/parse.ts
@@ -1,0 +1,3 @@
+export function parseMessage(raw: string): unknown {
+  return JSON.parse(raw);
+}
