+++ src/services/onMessage/types.ts
@@ -1,0 +1,4 @@
+export interface Message {
+  type: string;
+  payload?: unknown;
+}
