+++ src/utils/AGENTS.md
@@ -1,0 +1,11 @@
+# AGENTS
+
+Shared stateless helpers.
+
+- Keep functions pure and type-safe.
+- Place tests under `src/__tests__`.
+- After changes run:
+  ```bash
+  pre-commit run --files <files>
+  npm test -- --coverage
+  ```
