+++ src/utils/validation.ts
@@ -1,4 +1,4 @@
-import i18n from './i18n';
+import i18n from '../i18n';
 
 export const validateLight = (
   name: string,
